fix(IO): Move extractBinary to IO/Core/BinaryHelper
extractBinary is generic enough to be put in a general helper file.
Also adds a new method named arrayBufferToString that was excised out of
extractBinary.
