+++ Sources/IO/Core/index.js
@@ -1,8 +1,10 @@
+import BinaryHelper         from './BinaryHelper';
 import DataAccessHelper     from './DataAccessHelper';
 import vtkHttpDataSetReader from './HttpDataSetReader';
 import vtkHttpSceneLoader   from './HttpSceneLoader';
 
 export default {
+  BinaryHelper,
   DataAccessHelper,
   vtkHttpDataSetReader,
   vtkHttpSceneLoader,
