+++ Sources/IO/XML/XMLReader/index.js
@@ -18,48 +18,6 @@ function stringToXML(xmlStr) {
   return (new DOMParser()).parseFromString(xmlStr, 'application/xml');
 }
 
-/**
- * Extracts binary data out of a file bytearray given a prefix/suffix.
- */
-function extractBinary(arrayBuffer, prefixRegex, suffixRegex = null) {
-  // convert array buffer to string via fromCharCode so length is preserved
-  const byteArray = new Uint8Array(arrayBuffer);
-  const strArr = [];
-  for (let i = 0; i < byteArray.length; ++i) {
-    strArr[i] = String.fromCharCode(byteArray[i]);
-  }
-  const str = strArr.join('');
-
-  const prefixMatch = prefixRegex.exec(str);
-  if (!prefixMatch) {
-    return { text: str };
-  }
-
-  const dataStartIndex = prefixMatch.index + prefixMatch[0].length;
-  const strFirstHalf = str.substring(0, dataStartIndex);
-  let retVal = null;
-
-  const suffixMatch = suffixRegex ? suffixRegex.exec(str) : null;
-  if (suffixMatch) {
-    const strSecondHalf = str.substr(suffixMatch.index);
-    retVal = {
-      text: strFirstHalf + strSecondHalf,
-      binaryBuffer: arrayBuffer.slice(dataStartIndex, suffixMatch.index),
-    };
-  } else {
-    // no suffix, so just take all the data starting from dataStartIndex
-    retVal = {
-      text: strFirstHalf,
-      binaryBuffer: arrayBuffer.slice(dataStartIndex),
-    };
-  }
-
-  // TODO Maybe delete the internal ref to strArr from the match objs?
-  retVal.prefixMatch = prefixMatch;
-  retVal.suffixMatch = suffixMatch;
-  return retVal;
-}
-
 // ----------------------------------------------------------------------------
 
 const TYPED_ARRAY = {
